backend/requirements.txt
fastapi
uvicorn
pillow
opencv-python
numpy
transformers
torch   # required by transformers
